+++ src/assets/loader.js
@@ -1,6 +1,5 @@
 /* eslint-disable */
 require('file-loader?name=dictionary.json!../common/dictionary_en.json');
-require('babel-polyfill');
 require('whatwg-fetch');
 require('indexeddbshim');
 require('./asset-symbols/symbols.js');