+++ src/components/GPOSPanel/GPOSPanel.jsx
@@ -3,11 +3,12 @@ import {connect} from 'react-redux';
 import {bindActionCreators} from 'redux';
 import Translate from 'react-translate-component';
 import asset_utils from '../../common/asset_utils';
-import {HelpActions, DashboardPageActions} from '../../actions';
+import {HelpActions, DashboardPageActions, GPOSActions} from '../../actions';
 import {getTotalGposBalance} from '../../selectors/GPOSSelector';
 import {FormattedNumber} from 'react-intl';
 import {anchors} from '../Help/HelpModal';
 import AppUtils from '../../utility/AppUtils';
+import config from '../../../config/Config';
 
 class GPOSPanel extends Component {
   onClickHelpLearn = (e) => {
@@ -17,6 +18,10 @@ class GPOSPanel extends Component {
 
   componentDidMount() {
     this.props.fetchGposInfo();
+  }
+
+  openModal = () => {
+    this.props.toggleGposWizard(true);
   }
 
   renderGposStats = () => {
@@ -81,9 +86,20 @@ class GPOSPanel extends Component {
 
   render() {
     let {totalGpos} = this.props;
-    let stats = totalGpos && totalGpos > 0 ?
-      this.renderGposStats() : null;
-    let classModifier = totalGpos && totalGpos > 0 ? '' : '--no-stats';
+    let stats, classModifier;
+
+    // Render stats regardless of gpos balance.
+    stats = this.renderGposStats();
+    classModifier = '';
+
+    // If enabled in config, conditionally render the GPOS stats seciton.
+    if (config.gpos.conditionalStats) {
+      stats = totalGpos && totalGpos > 0 ?
+        this.renderGposStats() : null;
+    }
+
+    classModifier = totalGpos && totalGpos > 0 ? '' : '--no-stats';
+    classModifier = config.gpos.conditionalStats ? classModifier : `${classModifier} cond`;
     let btnString = totalGpos && totalGpos > 0 ? 'gpos.side.participate' : 'gpos.side.start';
 
     return (
@@ -109,7 +125,7 @@ class GPOSPanel extends Component {
         <button
           type='button'
           className={ `btn btn-content__head gpos-panel__btn${classModifier}` }
-          // onClick={ /* open gpos wizard */ }
+          onClick={ this.openModal }
         >
           <img className={ `gpos-panel__img-thumb${classModifier}` } src='images/thumb-up.png' alt='thumb'/>
           <Translate content={ btnString }/>
@@ -125,7 +141,7 @@ const mapStateToProps = (state) => {
   let gposInfo = state.dashboardPage.gposInfo;
   let gposReward = gposInfo && gposInfo.award && gposInfo.award.amount;
   let vestingFactor = gposInfo && gposInfo.vesting_factor;
-  let gposPerformance = AppUtils.trimNum((vestingFactor * 100 || 100), 2);
+  let gposPerformance = AppUtils.trimNum((vestingFactor * 100 || 0), 2);
 
   return {
     totalGpos: data.totalAmount,
@@ -138,6 +154,7 @@ const mapStateToProps = (state) => {
 const mapDispatchToProps = (dispatch) => bindActionCreators(
   {
     toggleHelpModal: HelpActions.toggleHelpAndScroll,
+    toggleGposWizard: GPOSActions.toggleGPOSWizardModal,
     fetchGposInfo: DashboardPageActions.getGposInfo
   },
   dispatch